remote_window/requirements.txt
tkinter
pillow
pyautoit